make MonoDetour reference bring in Mono.Cecil 0.11.5
MonoDetour.Bindings.Reorg needs 0.11.5 because MonoMod reorg uses it.

This change is because otherwise 0.11.4 or 0.11.5 would be available depending on references and the build system doesn't like that.
So people will get a newer Mono.Cecil version even if they have 0.11.4 at runtime, but apparently it's impossible to just ignore this issue.
This solution will be good enough for now, since 0.11.5 is mostly a bug fix release.