HookGen: change default namespace to 'Md'
While 'On' would be very good as a namespace, unfortunately C# is a great language,
and as such namespaces with zero public members will conflict with public types.

So, if you are an API and use MonoDetour.HookGen with 'On' namespace, anyone referencing
your API and using MonoMod's own HookGen will get conflicts due to the 'On' namespace
from the API having a namespace with the same name as a type from MonoMod's own HookGen
assemblies in the 'On' namespace.

There are multiple ways to go about this, such as generating MonoDetour's On' namespace
under the default namespace of the consumer. But that could come with its own headaches.

As such, I believe the safest solution is to just change the default HookGen namespace.